Correct bug in listening mode
I had assumed `net.Listen()` took as argument an address+port from which
to listen for connections (ie, *who* is making this request). This was
incorrect. It actually takes a *local* address+port onto which it should
 listen for incoming connections (inbound from *any* address).

This commit modifies the program's CLI arguments to allow users to
distinctly set the 'target' and 'listening' addresses (or to be more
precise, the 'listening port' since 'localhost' address is assumed).