Change Connection and TUI API to be function based
Previously, the Connection object exposed channels as its mechanism for
sending and retrieving messages. This API was nifty and showed a unique,
ideomatic feature of Go. Today I'm replacing that mechanism with a
function based API.

I'm doing this because I want to define Connection as an *interface*,
which requires it to be defined as a set of functions rather than
properties. The interface approach will allow for separate versions
of the Connection type: plaintext and encrypted.

This commit also changes the TUI API, which was previously composed of
one channel and one function mechanism. Leaving just one channel API
felt out of place, so I converted this too to a function.